# docs/openapi/brok-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brok API",
    "version": "1.0.0",
    "summary": "Brok public developer API",
    "description": "Machine-readable contract for Brok v1 chat, messages, search, models, and usage APIs. Streaming endpoints return Server-Sent Events and document event names with x-brok-sse-events."
  },
  "servers": [
    {
      "url": "https://www.brok.fyi",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BrokApiKey": []
    }
  ],
  "paths": {
...
    "/api/v1/messages": {
      "post": {
        "operationId": "createMessage",
        "summary": "Create an Anthropic-compatible message",
        "description": "Runs a Brok model with Anthropic-style message payloads. Requires code:write scope. Non-streaming billable responses support Idempotency-Key replay for 24 hours; streaming requests with the same key return 409 while in progress or after completion because stream bodies are not replayed.",
        "tags": ["Messages"],
        "x-brok-sse-events": [
          "content_block_start",
          "content_block_delta",
          "content_block_stop",
          "message_delta",
          "message_stop",
          "error"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message response or SSE stream",
            "headers": {
              "x-request-id": {
                "$ref": "#/components/headers/RequestId"
              },
              "idempotency-key": {
                "$ref": "#/components/headers/IdempotencyKey"
              },
              "idempotency-replayed": {
                "$ref": "#/components/headers/IdempotencyReplayed"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE stream using Anthropic-compatible event names."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BrokApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "brok_sk_*"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Optional client-generated key for billable POST retries. Keys are scoped to the API key, route, and exact request payload for 24 hours. Completed non-stream JSON responses replay with Idempotency-Replayed=true; streaming responses are not replayed.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      }
    },
    "headers": {
      "RequestId": {
        "description": "Request ID for support/debugging.",
        "schema": {
          "type": "string",
          "pattern": "^req_"
        }
      },
      "IdempotencyKey": {
        "description": "Echoes the accepted Idempotency-Key on new and replayed idempotent requests.",
        "schema": {
          "type": "string"
        }
      },
      "IdempotencyReplayed": {
        "description": "true when the response was replayed from the idempotency ledger; false for newly executed idempotent requests.",
        "schema": {
          "type": "string",
          "enum": ["true", "false"]
        }
      },
...
    },
    "responses": {
      "InvalidRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Inactive key, expired key, missing scope, model access, or feature access",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "IdempotencyConflict": {
        "description": "Idempotency-Key conflict, in-progress duplicate, failed prior attempt, or unsupported streaming replay",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "retry-after": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Temporary service, provider, rate-limit, or usage-ledger failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["type", "message"],
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Usage": {
        "type": "object",
        "properties": {
          "prompt_tokens": {
            "type": "integer"
          },
          "completion_tokens": {
            "type": "integer"
          },
          "total_tokens": {
            "type": "integer"
          },
          "search_queries": {
            "type": "integer"
          }
        },
        "additionalProperties": true
      },
...
      "MessagesRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": {
            "type": "string",
            "default": "brok-code"
          },
          "system": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            ]
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["role", "content"],
              "properties": {
                "role": {
                  "type": "string",
                  "enum": ["user", "assistant"]
                },
                "content": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  ]
                }
              },
              "additionalProperties": true
            }
          },
          "max_tokens": {
            "type": "integer",
            "minimum": 1
          },
          "temperature": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          },
          "stream": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": true
      },
      "MessagesResponse": {
        "type": "object",
        "required": ["id", "type", "role", "model", "content"],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "message"
          },
          "role": {
            "type": "string",
            "const": "assistant"
          },
          "model": {
            "type": "string"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "additionalProperties": true
      },
...
    }
  },
...
}